fix: resolve issue #81 - Move internal/wallet/ to internal/interface-adapters/wallet/
- Created new directory structure internal/interface-adapters/wallet/{btc,eth,xrp}
- Consolidated wallet interfaces (Keygener, Signer, Watcher) into interfaces.go
- Moved Bitcoin wallet implementation to btc/ with package name btc
- Moved Ethereum wallet implementation to eth/ with package name eth
- Moved XRP wallet implementation to xrp/ with package name xrp
- Updated all import paths throughout the codebase
- Updated type references and constructor calls in DI container and CLI adapters
- Added backward compatibility aliases in internal/wallet/ (temporary)
- Updated documentation (AGENTS.md, docs/directory_structure.md)
- Removed old wallet subdirectories (btcwallet, ethwallet, xrpwallet)
- All tests pass, linting passes, build succeeds

Closes #81
