+++ internal/interface-adapters/cli/watch/watch.go
@@ -13,10 +13,10 @@ import (
 	"github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/cli/watch/imports"
 	"github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/cli/watch/monitor"
 	"github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/cli/watch/send"
-	wallets "github.com/hiromaily/go-crypto-wallet/internal/wallet"
-	"github.com/hiromaily/go-crypto-wallet/internal/wallet/btcwallet"
-	"github.com/hiromaily/go-crypto-wallet/internal/wallet/ethwallet"
-	"github.com/hiromaily/go-crypto-wallet/internal/wallet/xrpwallet"
+	wallets "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet"
+	btcwallet "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet/btc"
+	ethwallet "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet/eth"
+	xrpwallet "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet/xrp"
 	"github.com/hiromaily/go-crypto-wallet/pkg/config"
 )
