+++ internal/interface-adapters/wallet/btc/watch.go
@@ -1,4 +1,4 @@
-package xrpwallet
+package btc
 
 import (
 	"context"
@@ -9,14 +9,15 @@ import (
 	domainCoin "github.com/hiromaily/go-crypto-wallet/internal/domain/coin"
 	domainTx "github.com/hiromaily/go-crypto-wallet/internal/domain/transaction"
 	domainWallet "github.com/hiromaily/go-crypto-wallet/internal/domain/wallet"
-	"github.com/hiromaily/go-crypto-wallet/internal/infrastructure/api/ripple"
-	"github.com/hiromaily/go-crypto-wallet/pkg/logger"
+	"github.com/hiromaily/go-crypto-wallet/internal/infrastructure/api/bitcoin"
+	"github.com/hiromaily/go-crypto-wallet/internal/infrastructure/storage/file/address"
 )
 
-// XRPWatch watch only wallet object
-type XRPWatch struct {
-	XRP                     ripple.Rippler
+// BTCWatch watch only wallet object
+type BTCWatch struct {
+	BTC                     bitcoin.Bitcoiner
 	dbConn                  *sql.DB
+	addrType                address.AddrType
 	wtype                   domainWallet.WalletType
 	createTxUseCase         watchusecase.CreateTransactionUseCase
 	monitorTxUseCase        watchusecase.MonitorTransactionUseCase
@@ -25,20 +26,22 @@ type XRPWatch struct {
 	createPaymentReqUseCase watchusecase.CreatePaymentRequestUseCase
 }
 
-// NewXRPWatch returns XRPWatch object
-func NewXRPWatch(
-	xrp ripple.Rippler,
+// NewBTCWatch returns Watch object
+func NewBTCWatch(
+	btc bitcoin.Bitcoiner,
 	dbConn *sql.DB,
+	addrType address.AddrType,
 	createTxUseCase watchusecase.CreateTransactionUseCase,
 	monitorTxUseCase watchusecase.MonitorTransactionUseCase,
 	sendTxUseCase watchusecase.SendTransactionUseCase,
 	importAddrUseCase watchusecase.ImportAddressUseCase,
 	createPaymentReqUseCase watchusecase.CreatePaymentRequestUseCase,
 	walletType domainWallet.WalletType,
-) *XRPWatch {
-	return &XRPWatch{
-		XRP:                     xrp,
+) *BTCWatch {
+	return &BTCWatch{
+		BTC:                     btc,
 		dbConn:                  dbConn,
+		addrType:                addrType,
 		wtype:                   walletType,
 		createTxUseCase:         createTxUseCase,
 		monitorTxUseCase:        monitorTxUseCase,
@@ -49,15 +52,15 @@ func NewXRPWatch(
 }
 
 // ImportAddress imports address
-func (w *XRPWatch) ImportAddress(fileName string, _ bool) error {
+func (w *BTCWatch) ImportAddress(fileName string, isRescan bool) error {
 	return w.importAddrUseCase.Execute(context.Background(), watchusecase.ImportAddressInput{
 		FileName: fileName,
-		Rescan:   false, // XRP doesn't support rescan
+		Rescan:   isRescan,
 	})
 }
 
 // createTx is a helper method to reduce code duplication across transaction creation methods
-func (w *XRPWatch) createTx(input watchusecase.CreateTransactionInput) (string, string, error) {
+func (w *BTCWatch) createTx(input watchusecase.CreateTransactionInput) (string, string, error) {
 	output, err := w.createTxUseCase.Execute(context.Background(), input)
 	if err != nil {
 		return "", "", err
@@ -66,46 +69,48 @@ func (w *XRPWatch) createTx(input watchusecase.CreateTransactionInput) (string, 
 }
 
 // CreateDepositTx creates deposit unsigned transaction
-func (w *XRPWatch) CreateDepositTx(_ float64) (string, string, error) {
+func (w *BTCWatch) CreateDepositTx(adjustmentFee float64) (string, string, error) {
 	return w.createTx(watchusecase.CreateTransactionInput{
-		ActionType: domainTx.ActionTypeDeposit.String(),
+		ActionType:    domainTx.ActionTypeDeposit.String(),
+		AdjustmentFee: adjustmentFee,
 	})
 }
 
 // CreatePaymentTx creates payment unsigned transaction
-func (w *XRPWatch) CreatePaymentTx(_ float64) (string, string, error) {
+func (w *BTCWatch) CreatePaymentTx(adjustmentFee float64) (string, string, error) {
 	return w.createTx(watchusecase.CreateTransactionInput{
-		ActionType: domainTx.ActionTypePayment.String(),
+		ActionType:    domainTx.ActionTypePayment.String(),
+		AdjustmentFee: adjustmentFee,
 	})
 }
 
 // CreateTransferTx creates transfer unsigned transaction
-func (w *XRPWatch) CreateTransferTx(
-	sender, receiver domainAccount.AccountType, floatAmount, _ float64,
+func (w *BTCWatch) CreateTransferTx(
+	sender, receiver domainAccount.AccountType, floatAmount, adjustmentFee float64,
 ) (string, string, error) {
 	return w.createTx(watchusecase.CreateTransactionInput{
 		ActionType:      domainTx.ActionTypeTransfer.String(),
 		SenderAccount:   sender,
 		ReceiverAccount: receiver,
 		Amount:          floatAmount,
+		AdjustmentFee:   adjustmentFee,
 	})
 }
 
 // UpdateTxStatus updates transaction status
-func (*XRPWatch) UpdateTxStatus() error {
-	logger.Info("no functionality for XRP")
-	return nil
+func (w *BTCWatch) UpdateTxStatus() error {
+	return w.monitorTxUseCase.UpdateTxStatus(context.Background())
 }
 
 // MonitorBalance monitors balance
-func (w *XRPWatch) MonitorBalance(confirmationNum uint64) error {
+func (w *BTCWatch) MonitorBalance(confirmationNum uint64) error {
 	return w.monitorTxUseCase.MonitorBalance(context.Background(), watchusecase.MonitorBalanceInput{
 		ConfirmationNum: confirmationNum,
 	})
 }
 
 // SendTx sends signed transaction
-func (w *XRPWatch) SendTx(filePath string) (string, error) {
+func (w *BTCWatch) SendTx(filePath string) (string, error) {
 	output, err := w.sendTxUseCase.Execute(context.Background(), watchusecase.SendTransactionInput{
 		FilePath: filePath,
 	})
@@ -116,27 +121,32 @@ func (w *XRPWatch) SendTx(filePath string) (string, error) {
 }
 
 // CreatePaymentRequest creates payment_request dummy data for development
-func (w *XRPWatch) CreatePaymentRequest() error {
+func (w *BTCWatch) CreatePaymentRequest() error {
 	amtList := []float64{
-		50,
-		100,
-		120,
-		130,
-		150,
+		0.00001,
+		0.00002,
+		0.000025,
+		0.000015,
+		0.00003,
 	}
+
 	return w.createPaymentReqUseCase.Execute(context.Background(), watchusecase.CreatePaymentRequestInput{
 		AmountList: amtList,
 	})
 }
 
 // Done should be called before exit
-func (w *XRPWatch) Done() {
+func (w *BTCWatch) Done() {
 	_ = w.dbConn.Close() // Best effort cleanup
-
-	_ = w.XRP.Close() // Best effort cleanup
+	w.BTC.Close()
 }
 
 // CoinTypeCode returns domainCoin.CoinTypeCode
-func (w *XRPWatch) CoinTypeCode() domainCoin.CoinTypeCode {
-	return w.XRP.CoinTypeCode()
+func (w *BTCWatch) CoinTypeCode() domainCoin.CoinTypeCode {
+	return w.BTC.CoinTypeCode()
 }
+
+// GetBTC gets btc
+// func (w *BTCWatch) GetBTC() bitcoin.Bitcoiner {
+//	return w.BTC
+//}
