+++ internal/interface-adapters/wallet/eth/keygen.go
@@ -1,4 +1,4 @@
-package ethwallet
+package eth
 
 import (
 	"context"
