+++ internal/interface-adapters/wallet/eth/sign.go
@@ -1,4 +1,4 @@
-package ethwallet
+package eth
 
 import (
 	"database/sql"
