+++ internal/interface-adapters/wallet/eth/watch.go
@@ -1,4 +1,4 @@
-package btcwallet
+package eth
 
 import (
 	"context"
@@ -9,15 +9,13 @@ import (
 	domainCoin "github.com/hiromaily/go-crypto-wallet/internal/domain/coin"
 	domainTx "github.com/hiromaily/go-crypto-wallet/internal/domain/transaction"
 	domainWallet "github.com/hiromaily/go-crypto-wallet/internal/domain/wallet"
-	"github.com/hiromaily/go-crypto-wallet/internal/infrastructure/api/bitcoin"
-	"github.com/hiromaily/go-crypto-wallet/internal/infrastructure/storage/file/address"
+	"github.com/hiromaily/go-crypto-wallet/internal/infrastructure/api/ethereum"
 )
 
-// BTCWatch watch only wallet object
-type BTCWatch struct {
-	BTC                     bitcoin.Bitcoiner
+// ETHWatch watch only wallet object
+type ETHWatch struct {
+	ETH                     ethereum.Ethereumer
 	dbConn                  *sql.DB
-	addrType                address.AddrType
 	wtype                   domainWallet.WalletType
 	createTxUseCase         watchusecase.CreateTransactionUseCase
 	monitorTxUseCase        watchusecase.MonitorTransactionUseCase
@@ -26,22 +24,20 @@ type BTCWatch struct {
 	createPaymentReqUseCase watchusecase.CreatePaymentRequestUseCase
 }
 
-// NewBTCWatch returns Watch object
-func NewBTCWatch(
-	btc bitcoin.Bitcoiner,
+// NewETHWatch returns ETHWatch object
+func NewETHWatch(
+	eth ethereum.Ethereumer,
 	dbConn *sql.DB,
-	addrType address.AddrType,
 	createTxUseCase watchusecase.CreateTransactionUseCase,
 	monitorTxUseCase watchusecase.MonitorTransactionUseCase,
 	sendTxUseCase watchusecase.SendTransactionUseCase,
 	importAddrUseCase watchusecase.ImportAddressUseCase,
 	createPaymentReqUseCase watchusecase.CreatePaymentRequestUseCase,
 	walletType domainWallet.WalletType,
-) *BTCWatch {
-	return &BTCWatch{
-		BTC:                     btc,
+) *ETHWatch {
+	return &ETHWatch{
+		ETH:                     eth,
 		dbConn:                  dbConn,
-		addrType:                addrType,
 		wtype:                   walletType,
 		createTxUseCase:         createTxUseCase,
 		monitorTxUseCase:        monitorTxUseCase,
@@ -52,15 +48,15 @@ func NewBTCWatch(
 }
 
 // ImportAddress imports address
-func (w *BTCWatch) ImportAddress(fileName string, isRescan bool) error {
+func (w *ETHWatch) ImportAddress(fileName string, _ bool) error {
 	return w.importAddrUseCase.Execute(context.Background(), watchusecase.ImportAddressInput{
 		FileName: fileName,
-		Rescan:   isRescan,
+		Rescan:   false, // ETH doesn't support rescan
 	})
 }
 
 // createTx is a helper method to reduce code duplication across transaction creation methods
-func (w *BTCWatch) createTx(input watchusecase.CreateTransactionInput) (string, string, error) {
+func (w *ETHWatch) createTx(input watchusecase.CreateTransactionInput) (string, string, error) {
 	output, err := w.createTxUseCase.Execute(context.Background(), input)
 	if err != nil {
 		return "", "", err
@@ -69,48 +65,45 @@ func (w *BTCWatch) createTx(input watchusecase.CreateTransactionInput) (string, 
 }
 
 // CreateDepositTx creates deposit unsigned transaction
-func (w *BTCWatch) CreateDepositTx(adjustmentFee float64) (string, string, error) {
+func (w *ETHWatch) CreateDepositTx(_ float64) (string, string, error) {
 	return w.createTx(watchusecase.CreateTransactionInput{
-		ActionType:    domainTx.ActionTypeDeposit.String(),
-		AdjustmentFee: adjustmentFee,
+		ActionType: domainTx.ActionTypeDeposit.String(),
 	})
 }
 
 // CreatePaymentTx creates payment unsigned transaction
-func (w *BTCWatch) CreatePaymentTx(adjustmentFee float64) (string, string, error) {
+func (w *ETHWatch) CreatePaymentTx(_ float64) (string, string, error) {
 	return w.createTx(watchusecase.CreateTransactionInput{
-		ActionType:    domainTx.ActionTypePayment.String(),
-		AdjustmentFee: adjustmentFee,
+		ActionType: domainTx.ActionTypePayment.String(),
 	})
 }
 
 // CreateTransferTx creates transfer unsigned transaction
-func (w *BTCWatch) CreateTransferTx(
-	sender, receiver domainAccount.AccountType, floatAmount, adjustmentFee float64,
+func (w *ETHWatch) CreateTransferTx(
+	sender, receiver domainAccount.AccountType, floatAmount, _ float64,
 ) (string, string, error) {
 	return w.createTx(watchusecase.CreateTransactionInput{
 		ActionType:      domainTx.ActionTypeTransfer.String(),
 		SenderAccount:   sender,
 		ReceiverAccount: receiver,
 		Amount:          floatAmount,
-		AdjustmentFee:   adjustmentFee,
 	})
 }
 
 // UpdateTxStatus updates transaction status
-func (w *BTCWatch) UpdateTxStatus() error {
+func (w *ETHWatch) UpdateTxStatus() error {
 	return w.monitorTxUseCase.UpdateTxStatus(context.Background())
 }
 
 // MonitorBalance monitors balance
-func (w *BTCWatch) MonitorBalance(confirmationNum uint64) error {
+func (w *ETHWatch) MonitorBalance(confirmationNum uint64) error {
 	return w.monitorTxUseCase.MonitorBalance(context.Background(), watchusecase.MonitorBalanceInput{
 		ConfirmationNum: confirmationNum,
 	})
 }
 
 // SendTx sends signed transaction
-func (w *BTCWatch) SendTx(filePath string) (string, error) {
+func (w *ETHWatch) SendTx(filePath string) (string, error) {
 	output, err := w.sendTxUseCase.Execute(context.Background(), watchusecase.SendTransactionInput{
 		FilePath: filePath,
 	})
@@ -121,32 +114,26 @@ func (w *BTCWatch) SendTx(filePath string) (string, error) {
 }
 
 // CreatePaymentRequest creates payment_request dummy data for development
-func (w *BTCWatch) CreatePaymentRequest() error {
+func (w *ETHWatch) CreatePaymentRequest() error {
 	amtList := []float64{
-		0.00001,
-		0.00002,
-		0.000025,
-		0.000015,
-		0.00003,
+		0.001,
+		0.002,
+		0.0025,
+		0.0015,
+		0.003,
 	}
-
 	return w.createPaymentReqUseCase.Execute(context.Background(), watchusecase.CreatePaymentRequestInput{
 		AmountList: amtList,
 	})
 }
 
 // Done should be called before exit
-func (w *BTCWatch) Done() {
+func (w *ETHWatch) Done() {
 	_ = w.dbConn.Close() // Best effort cleanup
-	w.BTC.Close()
+	w.ETH.Close()
 }
 
 // CoinTypeCode returns domainCoin.CoinTypeCode
-func (w *BTCWatch) CoinTypeCode() domainCoin.CoinTypeCode {
-	return w.BTC.CoinTypeCode()
+func (w *ETHWatch) CoinTypeCode() domainCoin.CoinTypeCode {
+	return w.ETH.CoinTypeCode()
 }
-
-// GetBTC gets btc
-// func (w *BTCWatch) GetBTC() bitcoin.Bitcoiner {
-//	return w.BTC
-//}
