+++ internal/interface-adapters/wallet/xrp/sign.go
@@ -1,0 +1,1 @@
+package xrp
