+++ internal/wallet/signer.go
@@ -1,16 +1,10 @@
 package wallets
 
 import (
-	domainKey "github.com/hiromaily/go-crypto-wallet/internal/domain/key"
+	wallet "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet"
 )
 
-// Signer is for signature wallet service interface
-type Signer interface {
-	GenerateSeed() ([]byte, error)
-	StoreSeed(strSeed string) ([]byte, error)
-	GenerateAuthKey(seed []byte, count uint32) ([]domainKey.WalletKey, error)
-	ImportPrivKey() error
-	ExportFullPubkey() (string, error)
-	SignTx(filePath string) (string, bool, string, error)
-	Done()
-}
+// Signer is a backward compatibility alias
+//
+// Deprecated: Use github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet.Signer instead
+type Signer = wallet.Signer
