+++ internal/wallet/watcher.go
@@ -1,22 +1,10 @@
 package wallets
 
 import (
-	domainAccount "github.com/hiromaily/go-crypto-wallet/internal/domain/account"
-	domainCoin "github.com/hiromaily/go-crypto-wallet/internal/domain/coin"
+	wallet "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet"
 )
 
-// Watcher is for watch only wallet service interface
-type Watcher interface {
-	ImportAddress(fileName string, isRescan bool) error
-	CreateDepositTx(adjustmentFee float64) (string, string, error)
-	CreatePaymentTx(adjustmentFee float64) (string, string, error)
-	CreateTransferTx(
-		sender, receiver domainAccount.AccountType, floatAmount, adjustmentFee float64,
-	) (string, string, error)
-	SendTx(filePath string) (string, error)
-	UpdateTxStatus() error
-	MonitorBalance(confirmationNum uint64) error
-	CreatePaymentRequest() error
-	Done()
-	CoinTypeCode() domainCoin.CoinTypeCode
-}
+// Watcher is a backward compatibility alias
+//
+// Deprecated: Use github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/wallet.Watcher instead
+type Watcher = wallet.Watcher
