--- internal/wallet/xrpwallet/sign.go
@@ -1,1 +1,0 @@
-package xrpwallet
